api modules rename
