+++ golem-worker-service/src/grpcapi/mod.rs
@@ -4,11 +4,11 @@ use golem_api_grpc::proto::golem::worker::worker_service_server::WorkerServiceSe
 use std::net::SocketAddr;
 use tonic::transport::{Error, Server};
 
-use crate::grpcapi::register_api_definition::GrpcApiDefinitionService;
+use crate::grpcapi::api_definition::GrpcApiDefinitionService;
 use crate::grpcapi::worker::WorkerGrpcApi;
 use crate::service::Services;
 
-mod register_api_definition;
+mod api_definition;
 mod worker;
 
 pub async fn start_grpc_server(addr: SocketAddr, services: &Services) -> Result<(), Error> {
